used more precise type in driver

+++ examples/vcsim01.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-    Datum<int> a(8), b(8);
+    Datum<uint8_t> a(8), b(8);
     a = 10; b = 20;
     vector<DatumBase*> inpv = {&a, &b};
     vcsim("vcsim01.vc", "addi", inpv);
